arm64: dts: 19821/19855: lower max brightness level to 2047
It turns out that brightness level 4095 corresponds exactly to the same
level as High Brightness Mode. That doesn't sound all that bad at first,
until you realize that the user might assume that the range of
brightness levels that they can set w/o HBM is considered to be
relatively safe. As such, we can end up with users setting max 4096
brightness all the time when using their phone, because in their minds,
it's still "safer" than HBM.

Not only that, but for our panel, when disabling HBM, the brightness
level that the panel returns to seems to be the equivalent of doing
`bl_lvl % 2048`. As such, if the backlight level was set somewere
around 87% in Android (corresponding to around ~2048 in the bl_lvl
value), then turning off HBM (or exiting out of udfps fp auth) results
in the backlight drastically dropping until the user readjusts the
backlight again.

Now, the latter problem (brightness wraparound) could be solved by just
immediately setting the bl_lvl again after turning off HBM, but that
leaves the former problem still. As such, it makes more sense to just
lower the max brightness range to 2047. HBM will still work as expected.

Change-Id: I30a96aa0b25d2694be0c678afa136957fee2f95c